+++ Trady.Analysis/Indicator/AccumulationDistributionLine.cs
@@ -5,7 +5,7 @@ using Trady.Core;
 
 namespace Trady.Analysis.Indicator
 {
-    public partial class AccumulationDistributionLine : CummulativeAnalyzableBase<(decimal High, decimal Low, decimal Close, decimal Volume), decimal?>
+    public partial class AccumulationDistributionLine : CumulativeAnalyzableBase<(decimal High, decimal Low, decimal Close, decimal Volume), decimal?>
     {
         public AccumulationDistributionLine(IList<Candle> candles) :
             this(candles.Select(i => (i.High, i.Low, i.Close, i.Volume)).ToList())
@@ -22,7 +22,7 @@ namespace Trady.Analysis.Indicator
 
         protected override decimal? ComputeInitialValue(int index) => Inputs[index].Volume;
 
-        protected override decimal? ComputeCummulativeValue(int index, decimal? prevOutput)
+        protected override decimal? ComputeCumulativeValue(int index, decimal? prevOutput)
         {
             var input = Inputs[index];
             var prevInput = Inputs[index - 1];
